requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-filter==2.4.0